Tweak xml tweaking of android build process
