+++ packages/cordova/package.json
@@ -77,7 +77,7 @@
     },
     "scripts": {
         "patch-ios": "sed -i '' s/UIApplication\\.openSettingsURLString/UIApplicationOpenSettingsURLString/g ./plugins/cordova-plugin-qrscanner/src/ios/QRScanner.swift ./platforms/ios/Padloc/Plugins/cordova-plugin-qrscanner/QRScanner.swift",
-        "prepare-build": "webpack && node ./update-config-xml.js && cordova prepare && node ./update-config-xml.js",
+        "prepare-build": "webpack && node ./update-config-xml.js && cordova prepare && node ./update-android-xmls.js",
         "start:android": "npm run prepare-build && cordova run android",
         "start:ios": "npm run prepare-build && npm run patch-ios && cordova run ios",
         "build": "npm run prepare-build && cordova build",
